fix: use Vulkan as the default (and recommended) backend
As of at least June of 2023, the upstream code introduced a change that causes
games to lose input after losing focus for the first time (until the next time
the user launches dolphin-emu) in OpenGL render mode. The upstream is yet to
identify the root cause of this issue (citing the lack of Linux hardware on
hand) - so until we find a solution on our own, demote OpenGL as the default
rendering backend to avoid confusion.

Promote Vulkan as the preferred rendering backend since it is probably readily
available on hardware that supports OpenGL 3.0 (which the emulator requires).

Fingers crossed for that being the case.

See:

https://bugs.dolphin-emu.org/issues/13354
dolphin-emu@c63f0f3#diff-4758964fe112db861ede53e23f9660201a2799dcad3428f3fcfdc137504e4b39